Login component enhancement
Used form and 'form' attributes of default input and button HTML elements to receive email and password information in Login component on submit button click.
