+++ raaghu-elements/.storybook/main.js
@@ -4,7 +4,7 @@ module.exports = {
     "../src/**/*.stories.@(js|jsx|ts|tsx)",
     "../src/**/**/*.stories.mdx",
     "../src/**/**/*.stories.@(js|jsx|ts|tsx)",
-    "../../raaghu-mfe/rds_component/**/**/*.stories.@(js|jsx|ts|tsx)"
+    "../../raaghu-mfe/rds_components/**/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
     "@storybook/addon-links",
