+++ raaghu-mfe/rds_components/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "rds_component",
+  "name": "rds_components",
   "version": "1.0.0",
   "description": "",
   "main": "index.ts",
