+++ raaghu-mfe/rds_pages/dashboard/webpack.config.js
@@ -63,7 +63,7 @@ module.exports = (env, argv) => {
         "./Dashboard": "./src/Dashboard/Dashboard",
       },
         remotes: {
-          rds_component: isProduction ? process.env.PROD_APP1 : process.env.DEV_APP1,
+          rds_components: isProduction ? process.env.PROD_APP1 : process.env.DEV_APP1,
           
         },
         shared: {
