+++ raaghu-mfe/rds_pages/login/.env
@@ -1,4 +1,4 @@
-DEV_APP1="rds_component@http://localhost:8001/remoteEntry.js"
+DEV_APP1="rds_components@http://localhost:8001/remoteEntry.js"
 
 
-PROD_APP1="rds_component@http://ogz-microfrontend-app-1.s3-website.eu-central-1.amazonaws.com/remoteEntry.js"
+PROD_APP1="rds_components@http://ogz-microfrontend-app-1.s3-website.eu-central-1.amazonaws.com/remoteEntry.js"
