+++ raaghu-mfe/rds_pages/login/src/remote.d.ts
@@ -1,6 +1,6 @@
 ///<reference types="react" />
 
-declare module "rds_component/Rdscomplogin" {
+declare module "rds_components/Rdscomplogin" {
 	const LoginComponent: React.ComponentType;
 
 	export default LoginComponent;
